requirements.txt
streamlit==1.40.0
streamlit-authenticator==0.4.1
pandas==2.2.3
numpy==2.1.3
yfinance==0.2.50
TA-Lib==0.4.32
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
pyyaml==6.0.1
psutil==5.9.8
openai==1.54.3
plotly==5.22.0

